Yoctopuce specific code (svn rev 69229)

+++ py/misc.h
@@ -461,9 +461,9 @@ static inline uint32_t mp_clz_mpi(mp_int_t x) {
 
     // ugly, but should compile to single intrinsic unless O0 is set
     if (mp_check(sizeof(mp_int_t) == sizeof(long))) {
-        return mp_clzl((unsigned long)x);
+        return (uint32_t)mp_clzl((unsigned long)x);
     } else {
-        return mp_clzll((unsigned long long)x);
+        return (uint32_t)mp_clzll((unsigned long long)x);
     }
     #endif
 }
